+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/components/PodcastCover.kt
@@ -24,7 +24,8 @@ fun PodcastCover(
     cornerRadius: Dp? = null,
 ) {
     val cornerRadiusSize = cornerRadius ?: if (coverSize == CoverSize.SMALL) 4.dp else 8.dp
-    PodcastImage(
+    @Suppress("DEPRECATION")
+    PodcastImageDeprecated(
         uuid = uuid,
         elevation = if (coverSize == CoverSize.SMALL) 4.dp else 8.dp,
         cornerSize = cornerRadiusSize,