+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/ui/podcast/PodcastScreen.kt
@@ -23,7 +23,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.wear.compose.foundation.lazy.items
 import androidx.wear.compose.material.MaterialTheme
 import androidx.wear.compose.material.Text
-import au.com.shiftyjelly.pocketcasts.compose.components.PodcastImage
+import au.com.shiftyjelly.pocketcasts.compose.components.PodcastImageDeprecated
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
@@ -87,7 +87,8 @@ private fun Content(
         ) {
             item { Spacer(Modifier.height(4.dp)) }
             item {
-                PodcastImage(
+                @Suppress("DEPRECATION")
+                PodcastImageDeprecated(
                     uuid = podcast.uuid,
                     modifier = Modifier.size(PodcastScreen.podcastImageSize),
                 )